Fix various uses of numeric_cast in UI bits
Apart from QtUI bits, this addresses use of uncaught numeric cast
exceptions in message count handling, the spell parser and the MacOS
idle querier.

The WindowsServicePrincipalName logic previously had an issue where
using ports from 32768 onwards would result in a bad_numeric_cast
exception to be thrown. This has been addressed at the same time as the
uncaught exceptions, and all ports should work now.

The tags file has been extended to ignore more files.

+++ SwifTools/Dock/Dock.h
@@ -1,10 +1,12 @@
 /*
- * Copyright (c) 2010 Isode Limited.
+ * Copyright (c) 2010-2018 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
 
 #pragma once
+
+#include <cstddef>
 
 namespace Swift {
 
@@ -13,6 +15,6 @@ namespace Swift {
         public:
             virtual ~Dock();
 
-            virtual void setNumberOfPendingMessages(int i) = 0;
+            virtual void setNumberOfPendingMessages(size_t i) = 0;
     };
 }
