+++ SwifTools/Dock/MacOSXDock.mm
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015-2016 Isode Limited.
+ * Copyright (c) 2015-2018 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
@@ -18,7 +18,7 @@ namespace Swift {
 MacOSXDock::MacOSXDock(CocoaApplication*) {
 }
 
-void MacOSXDock::setNumberOfPendingMessages(int i) {
+void MacOSXDock::setNumberOfPendingMessages(size_t i) {
     std::string label(i > 0 ? boost::lexical_cast<std::string>(i) : "");
     NSString *labelString = [[NSString alloc] initWithUTF8String: label.c_str()];
     [[NSApp dockTile] setBadgeLabel: labelString];
