+++ SwifTools/Dock/NullDock.h
@@ -1,10 +1,12 @@
 /*
- * Copyright (c) 2010 Isode Limited.
+ * Copyright (c) 2010-2018 Isode Limited.
  * All rights reserved.
  * See the COPYING file for more information.
  */
 
 #pragma once
+
+#include <cstddef>
 
 #include <SwifTools/Dock/Dock.h>
 
@@ -13,7 +15,7 @@ namespace Swift {
         public:
             NullDock() {}
 
-            virtual void setNumberOfPendingMessages(int) {
+            virtual void setNumberOfPendingMessages(size_t) {
             }
     };
 }
